+++ internal/probe/egress/flow/bpf/connect.bpf.c
@@ -18,6 +18,7 @@
 #include "vmlinux.h"
 #include <bpf/bpf_helpers.h>
 #include <bpf/bpf_tracing.h>
+#include <bpf/bpf_endian.h>
 
 char __license[] SEC("license") = "EUPL-1.2";